savefrom/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
click==8.1.7
exceptiongroup==1.2.1
fastapi==0.110.2
h11==0.14.0
idna==3.7
pydantic==2.7.0
pydantic_core==2.18.1
sniffio==1.3.1
starlette==0.37.2
typing_extensions==4.11.0
uvicorn==0.29.0
annotated-types==0.6.0
anyio==4.3.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.5
exceptiongroup==1.2.1
fastapi==0.110.2
h11==0.14.0
htmlmin==0.1.12
idna==3.7
Jinja2==3.1.3
MarkupSafe==2.1.5
pycparser==2.22
pydantic==2.7.0
pydantic_core==2.18.1
python-dotenv==1.0.1
requests==2.31.0
sniffio==1.3.1
starlette==0.37.2
typing_extensions==4.11.0
urllib3==2.2.1
uvicorn==0.29.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.5
exceptiongroup==1.2.1
fastapi==0.110.2
frozenlist==1.4.1
h11==0.14.0
htmlmin==0.1.12
httpcore==1.0.5
httpx==0.27.0
idna==3.7
Jinja2==3.1.3
MarkupSafe==2.1.5
multidict==6.0.5
Naked==0.1.32
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.7.0
pydantic_core==2.18.1
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.31.0
shellescape==3.8.1
sniffio==1.3.1
starlette==0.37.2
typing_extensions==4.11.0
urllib3==1.26.6
uvicorn==0.29.0
yarl==1.9.4
